# Speed up QTree