This makes an algorithmic improvement, and uses sumOption to get pretty big wins in performance which should be enough for map/reduce applications. There are two tricks:

1) we don't need to scan the whole tree if the minCount is 1 or less, which is common, especially for smaller QTrees, which come up when you just add a few together which would be done in map-side aggregation.

2) in sumOption, we can wait to compress after we have added a number of items together. This can be controlled by the user but defaults to 25, which seems to be big enough to give a win in one of the benchmark cases, but also not so huge that I am nervous the QTree is going to get out of hand in size.

Also, I removed the `Monoid[A]` from the QTree instance, which was probably being serialized when people use Kryo, which is no good. Also, it really belongs in the Semigroup (which I think could be a Monoid if we needed, just make a zero with a zero count).

related to #385 

/cc @avibryant 

Benchmark results:
```
Old code:

[info] 0% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=5, numElements=100} 984899.57 ns;
σ=57042.22 ns @ 10 trials
[info] 17% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=10, numElements=100} 1083250.54
ns; σ=5840.94 ns @ 3 trials
[info] 33% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=12, numElements=100} 1137581.99
ns; σ=108320.10 ns @ 10 trials
[info] 50% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=5, numElements=10000} 61440033.33
ns; σ=2959694.26 ns @ 10 trials
[info] 67% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=10, numElements=10000}
1150803500.00 ns; σ=10953173.74 ns @ 4 trials
[info] 83% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=12, numElements=10000}
795943000.00 ns; σ=5517007.01 ns @ 3 trials
[info] 
[info] numElements depthK      us linear runtime
[info]         100      5     985 =
[info]         100     10    1083 =
[info]         100     12    1138 =
[info]       10000      5   61440 =
[info]       10000     10 1150804 ==============================
[info]       10000     12  795943 ====================
[info] 
[info] vm: java
[info] trial: 0
[info] benchmark: SumOption

New code: 
[info]  0% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=5, numElements=100} 131886.03 ns;
σ=189.18 ns @ 3 trials
[info] 17% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=10, numElements=100} 117297.16 ns;
σ=163.64 ns @ 3 trials
[info] 33% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=12, numElements=100} 118775.84 ns;
σ=637.37 ns @ 3 trials
[info] 50% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=5, numElements=10000} 15500888.37
ns; σ=428739.48 ns @ 10 trials
[info] 67% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=10, numElements=10000} 39304494.17
ns; σ=1701883.67 ns @ 10 trials
[info] 83% Scenario{vm=java, trial=0, benchmark=SumOption, depthK=12, numElements=10000} 23124595.24
ns; σ=273888.75 ns @ 10 trials
[info] 
[info] numElements depthK    us linear runtime
[info]         100      5   132 =
[info]         100     10   117 =
[info]         100     12   119 =
[info]       10000      5 15501 ===========
[info]       10000     10 39304 ==============================
[info]       10000     12 23125 =================
[info] 
[info] vm: java
[info] trial: 0
[info] benchmark: SumOption

if we compress after each sum (just looking at the other optimizations):

[info] numElements depthK     us linear runtime
[info]         100      5    227 =
[info]         100     10    119 =
[info]         100     12    125 =
[info]       10000      5  37222 ===
[info]       10000     10 357207 ==============================
[info]       10000     12  61052 =====
```