@Gabriel439 2015-04-16
Why is the test for `HyperLogLog` failing even though this only modifies `QTree`?