@johnynek 2015-05-05
Related to #429